+++ packages/common/hangul/src/index.ts
@@ -1,5 +1,6 @@
 /** @tossdocs-ignore */
-export * from './disassemble';
-export * from './hangulIncludes';
 export * from './chosungIncludes';
+export * from './disassembleHangul';
+export * from './disassembleHangulToGroups';
+export * from './hangulIncludes';
 export * from './josa';
