+++ packages/common/hangul/src/josa.ko.md
@@ -1,0 +1,25 @@
+---
+title: josa
+---
+
+# josa
+
+한글 문자열에 `'이/가'`, `'을/를'`, `'은/는'`, `'으로/로'`, `'와/과'`, `'이나/나'`, `'이에/에'` 와 같은 조사를 붙여줍니다.
+
+```typescript
+function josa(
+  // 조사를 붙일 한글 문자열
+  word: string,
+  // 붙일 조사
+  josa: '이/가' | '을/를' | '은/는' | '으로/로' | '와/과' | '이나/나' | '이에/에'
+): string;
+```
+
+## Examples
+
+```typescript
+josa('샴푸', '이/가'); // '샴푸가'
+josa('칫솔', '이/가'); // '칫솔이'
+josa('바깥', '으로/로'); // '바깥으로'
+josa('내부', '으로/로'); // '내부로'
+```
